AreaRatio as float

+++ Pacmio/Source/Modules/BarTable/Analysis/Basic/ATR.cs
@@ -95,7 +95,7 @@ namespace Pacmio
 
         public string AreaName { get; protected set; }
 
-        public int AreaRatio { get; set; } = 8;
+        public float AreaRatio { get; set; } = 8;
 
         public int AreaOrder { get; set; } = 0;
