+++ Pacmio/Source/Modules/BarTable/Analysis/Indicators/WaveTrend.cs
@@ -179,7 +179,7 @@ namespace Pacmio // Can be derived from SMA
 
         public string AreaName { get; }
 
-        public int AreaRatio { get; set; } = 10;
+        public float AreaRatio { get; set; } = 10;
 
         public int AreaOrder { get; set; } = 0;
